collect asfreq tests
